+++ kubejs/server_scripts/NTC/base/modrecipes/applied_energistics2/creative_storage_cell.js
@@ -1,0 +1,23 @@
+events.listen('recipes', function (event) {
+  event.recipes.extendedcrafting.shaped_table('appliedenergistics2:creative_storage_cell', [
+    'NWWWWWWWN',
+    'WENNYNNEW',
+    'WNAYXYANW',
+    'WNYSYSYNW',
+    'WYXYZYXYW',
+    'WNYSYSYNW',
+    'WNAYXYANW',
+    'WENNYNNEW',
+    'NPPPPPPPN'
+  ], {
+    E: 'avaritia:endest_pearl',
+    N: 'avaritia:infinity_ingot',
+    P: 'extendedcrafting:the_ultimate_component',
+    S: 'extendedcrafting:ultimate_singularity',
+    A: 'projectexpansion:final_star',
+    W: 'betternether:quartz_glass_framed',
+    X: 'compressium:lapis_9',
+    Y: 'compressium:redstone_9',
+    Z: 'appliedenergistics2:empty_storage_cell'
+  }).id(`djefnukem:appliedenergistics2/creative_storage_cell`)
+})