+++ kubejs/server_scripts/NTC/base/modrecipes/tconstruct/creative_slots.js
@@ -1,0 +1,86 @@
+events.listen('recipes', function (event) {
+  event.recipes.extendedcrafting.shaped_table(Item.of('tconstruct:creative_slot', '{slot:"abilities"}'), [
+    'OOOTSTOOO',
+    'OOOIEIOOO',
+    'OOYIYIYOO',
+    'TIICFCIIT',
+    'SEYFAFYES',
+    'TIICFCIIT',
+    'OOYIYIYOO',
+    'OOOIEIOOO',
+    'OOOTSTOOO'
+  ], {
+    O: 'minecraft:air',
+    S: 'extendedcrafting:ultimate_singularity',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    A: 'projectexpansion:final_star',
+    E: 'avaritia:endest_pearl',
+    Y: 'avaritia:infinity_catalyst',
+    C: 'kubejs:ultimate_conglomerate',
+    F: 'miniutilities:the_final_opinium_core'
+  }).id(`djefnukem:tconstruct/creative_ability`)
+  event.recipes.extendedcrafting.shaped_table(Item.of('tconstruct:creative_slot', '{slot:"souls"}'), [
+    'OOOIYIOOO',
+    'OOOIYIOOO',
+    'OOICFCIOO',
+    'SITYFYTIS',
+    'TTEEAEETT',
+    'SITYFYTIS',
+    'OOICFCIOO',
+    'OOOIYIOOO',
+    'OOOIYIOOO'
+  ], {
+    O: 'minecraft:air',
+    S: 'extendedcrafting:ultimate_singularity',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    A: 'projectexpansion:final_star',
+    E: 'avaritia:endest_pearl',
+    Y: 'avaritia:infinity_catalyst',
+    C: 'kubejs:ultimate_conglomerate',
+    F: 'miniutilities:the_final_opinium_core'
+  }).id(`djefnukem:tconstruct/creative_soul_slot`)
+  event.recipes.extendedcrafting.shaped_table(Item.of('tconstruct:creative_slot', '{slot:"upgrades"}'), [
+    'OOFTOOOOO',
+    'OOFTYYYYT',
+    'OOFTYYYYT',
+    'OOFTOOOOO',
+    'OOOOOOOOO',
+    'IIIIIIIII',
+    'IIIEAEIIO',
+    'OSCEIECSO',
+    'OSCTITCSO'
+  ], {
+    O: 'minecraft:air',
+    S: 'extendedcrafting:ultimate_singularity',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    A: 'projectexpansion:final_star',
+    E: 'avaritia:endest_pearl',
+    Y: 'avaritia:infinity_catalyst',
+    C: 'kubejs:ultimate_conglomerate',
+    F: 'miniutilities:the_final_opinium_core'
+  }).id(`djefnukem:tconstruct/creative_upgrade`)
+  event.recipes.extendedcrafting.shaped_table(Item.of('tconstruct:creative_slot', '{slot:"defense"}'), [
+    'OOOOOOOOO',
+    'OOFOOOFOO',
+    'OOITTTIOO',
+    'YEIIIIIEY',
+    'YEISFSIEY',
+    'OYISFSIYO',
+    'OYIIIIIYO',
+    'OOCCACCOO',
+    'OOTTTTTOO'
+  ], {
+    O: 'minecraft:air',
+    S: 'extendedcrafting:ultimate_singularity',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    A: 'projectexpansion:final_star',
+    E: 'avaritia:endest_pearl',
+    Y: 'avaritia:infinity_catalyst',
+    C: 'kubejs:ultimate_conglomerate',
+    F: 'miniutilities:the_final_opinium_core'
+  }).id(`djefnukem:tconstruct/creative_defense_slot`)
+})