+++ kubejs/server_scripts/NTC/base/modrecipes/thermal/creative_augments.js
@@ -1,0 +1,90 @@
+events.listen('recipes', function (event) {
+  event.recipes.extendedcrafting.shaped_table('thermal:machine_efficiency_creative_augment', [
+    'YIOTLTOIY',
+    'IYOOLOOYI',
+    'OOSLLLSOO',
+    'TOMBCBMOT',
+    'MMMCACMMM',
+    'TOMBCBMOT',
+    'OOSLLLSOO',
+    'IYOOLOOYI',
+    'YIOTLTOIY'
+  ], {
+    O: 'minecraft:air',
+    Y: 'avaritia:infinity_catalyst',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    B: 'avaritia:infinity_block',
+    C: 'kubejs:ultimate_conglomerate',
+    A: 'projectexpansion:final_star',
+    S: 'extendedcrafting:ultimate_singularity',
+    L: 'thermal:machine_efficiency_augment',
+    M: 'thermal:machine_speed_augment'
+  }).id(`djefnukem:thermal/creative_flux_efficiency`)
+  event.recipes.extendedcrafting.shaped_table('thermal:rf_coil_creative_augment', [
+    'YIOTFTOIY',
+    'IYOOHOOYI',
+    'OOSFHFSOO',
+    'TOGBCBGOT',
+    'GHHCACHHG',
+    'TOGBCBGOT',
+    'OOSFHFSOO',
+    'IYOOHOOYI',
+    'YIOTFTOIY'
+  ], {
+    O: 'minecraft:air',
+    Y: 'avaritia:infinity_catalyst',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    B: 'avaritia:infinity_block',
+    C: 'kubejs:ultimate_conglomerate',
+    A: 'projectexpansion:final_star',
+    S: 'extendedcrafting:ultimate_singularity',
+	H: 'thermal:rf_coil_augment',
+    G: 'thermal:rf_coil_xfer_augment',
+    F: 'thermal:rf_coil_storage_augment'
+  }).id(`djefnukem:thermal/creative_rf_coil`)
+  event.recipes.extendedcrafting.shaped_table('thermal:fluid_tank_creative_augment', [
+    'YIOTDTOIY',
+    'IYOODOOYI',
+    'OOSDDDSOO',
+    'TODBCBDOT',
+    'DDDCACDDD',
+    'TODBCBDOT',
+    'OOSDDDSOO',
+    'IYOODOOYI',
+    'YIOTDTOIY'
+  ], {
+    O: 'minecraft:air',
+    Y: 'avaritia:infinity_catalyst',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    B: 'avaritia:infinity_block',
+    C: 'kubejs:ultimate_conglomerate',
+    A: 'projectexpansion:final_star',
+    S: 'extendedcrafting:ultimate_singularity',
+    D: 'thermal:fluid_tank_augment'
+  }).id(`djefnukem:thermal/creative_tank_construction`)
+  event.recipes.extendedcrafting.shaped_table('thermal:machine_catalyst_creative_augment', [
+    'YIOTDTOIY',
+    'IYOODOOYI',
+    'OOSDDDSOO',
+    'TOXBCBXOT',
+    'XXXCACXXX',
+    'TOXBCBXOT',
+    'OOSDDDSOO',
+    'IYOODOOYI',
+    'YIOTDTOIY'
+  ], {
+    O: 'minecraft:air',
+    Y: 'avaritia:infinity_catalyst',
+    I: 'extendedcrafting:the_ultimate_ingot',
+    T: 'extendedcrafting:the_ultimate_catalyst',
+    B: 'avaritia:infinity_block',
+    C: 'kubejs:ultimate_conglomerate',
+    A: 'projectexpansion:final_star',
+    S: 'extendedcrafting:ultimate_singularity',
+    X: 'thermal:machine_output_augment',
+    D: 'thermal:machine_catalyst_augment'
+  }).id(`djefnukem:thermal/creative_reclamation_chamber`)
+})